Move ProductRelease.SupportLevel to constants module
